+++ Classes/BackEnd/Module.php
@@ -1,5 +1,6 @@
 <?php
 
+use OliverKlee\Phpunit\Service\LibraryLoader;
 use TYPO3\CMS\Backend\Module\BaseScriptClass;
 use TYPO3\CMS\Backend\Template\DocumentTemplate;
 use TYPO3\CMS\Backend\Utility\BackendUtility;
@@ -211,7 +212,7 @@ class Tx_Phpunit_BackEnd_Module extends BaseScriptClass
      */
     public function main()
     {
-        require_once PATH_site . 'typo3conf/ext/phpunit/Resources/Private/Libraries/phpunit-library.phar';
+        LibraryLoader::includeAll();
 
         $this->doc = GeneralUtility::makeInstance(DocumentTemplate::class);
         $this->doc->backPath = $GLOBALS['BACK_PATH'];